modelo y tablas pedidos y productos

+++ db/migrate/20170701041105_create_hojas.rb
@@ -1,7 +1,7 @@
 class CreateHojas < ActiveRecord::Migration[5.1]
   def change
     create_table :hojas do |t|
-      t.integer :porcion
+      t.float :porcion
 
       t.timestamps
     end
